# specs/002-jupyterhub-notebook/contracts/api.yaml
openapi: 3.0.3
info:
  title: ML Platform API - Analysis, Workspace & Notebooks
  description: >
    REST API for feature 002-jupyterhub-notebook.
    Covers analysis management, workspace lifecycle management
    (launch, status, terminate) scoped per analysis, and JupyterLab
    URL retrieval for iframe embedding.
  version: 0.2.0

servers:
  - url: http://localhost:8080/api/v1
    description: Local development

security:
  - bearerAuth: []

paths:
  /analyses:
    post:
      summary: Create a new analysis
      description: >
        Creates a new analysis for the authenticated user. Each analysis
        is a named project context that scopes its own workspace and
        MLflow experiments.
      operationId: createAnalysis
      tags: [Analysis]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAnalysisRequest'
      responses:
        '201':
          description: Analysis created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AnalysisInfo'
        '401':
          description: Invalid or missing token
        '409':
          description: Analysis with this name already exists for the user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    get:
      summary: List user's analyses
      description: >
        Returns all analyses belonging to the authenticated user,
        ordered by creation time (newest first).
      operationId: listAnalyses
      tags: [Analysis]
      responses:
        '200':
          description: List of analyses
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AnalysisInfo'
        '401':
          description: Invalid or missing token

  /analyses/{analysisId}:
    get:
      summary: Get analysis details
      description: >
        Returns details of a specific analysis. Verifies the analysis
        belongs to the authenticated user.
      operationId: getAnalysis
      tags: [Analysis]
      parameters:
        - $ref: '#/components/parameters/analysisId'
      responses:
        '200':
          description: Analysis details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AnalysisInfo'
        '401':
          description: Invalid or missing token
        '404':
          description: Analysis not found or not owned by user

    delete:
      summary: Delete an analysis
      description: >
        Deletes an analysis and its associated resources. The analysis
        workspace must be stopped before deletion.
      operationId: deleteAnalysis
      tags: [Analysis]
      parameters:
        - $ref: '#/components/parameters/analysisId'
      responses:
        '204':
          description: Analysis deleted
        '401':
          description: Invalid or missing token
        '404':
          description: Analysis not found or not owned by user
        '409':
          description: Analysis has an active workspace that must be stopped first

  /workspaces/profiles:
    get:
      summary: List available compute profiles
      description: >
        Returns the list of compute profiles available for launching
        notebook servers. For MVP, only the "Exploratory" profile
        is returned. Profiles are not analysis-specific.
      operationId: getProfiles
      tags: [Workspace]
      responses:
        '200':
          description: List of compute profiles
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ComputeProfile'
        '401':
          description: Invalid or missing token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /analyses/{analysisId}/workspaces:
    post:
      summary: Launch a notebook workspace
      description: >
        Creates and starts a new notebook server (JupyterHub named
        server) for the specified analysis. If the analysis already has
        an active workspace, returns 409 Conflict. The server is
        launched with the specified compute profile (defaults to
        "Exploratory").
      operationId: launchWorkspace
      tags: [Workspace]
      parameters:
        - $ref: '#/components/parameters/analysisId'
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LaunchWorkspaceRequest'
      responses:
        '202':
          description: Workspace launch initiated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkspaceStatus'
        '401':
          description: Invalid or missing token
        '404':
          description: Analysis not found
        '409':
          description: Analysis already has an active workspace
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    get:
      summary: Get current workspace status
      description: >
        Returns the status of the workspace for the specified analysis.
        If no workspace exists, returns the status as "STOPPED" with
        no additional details.
      operationId: getWorkspaceStatus
      tags: [Workspace]
      parameters:
        - $ref: '#/components/parameters/analysisId'
      responses:
        '200':
          description: Current workspace status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkspaceStatus'
        '401':
          description: Invalid or missing token

    delete:
      summary: Terminate the workspace
      description: >
        Stops and terminates the analysis's active notebook server.
        User data in the persistent volume is preserved.
      operationId: terminateWorkspace
      tags: [Workspace]
      parameters:
        - $ref: '#/components/parameters/analysisId'
      responses:
        '204':
          description: Workspace termination initiated
        '401':
          description: Invalid or missing token
        '404':
          description: No active workspace found for this analysis

  /analyses/{analysisId}/workspaces/url:
    get:
      summary: Get JupyterLab URL for iframe embedding
      description: >
        Returns the JupyterLab URL for the analysis's running workspace.
        The URL includes the named server path segment. Used by the
        frontend to populate the iframe src attribute. Returns 404 if
        no workspace is running.
      operationId: getWorkspaceUrl
      tags: [Workspace]
      parameters:
        - $ref: '#/components/parameters/analysisId'
      responses:
        '200':
          description: JupyterLab URL
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkspaceUrl'
        '401':
          description: Invalid or missing token
        '404':
          description: No running workspace for this analysis
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        Keycloak-issued JWT from feature 001 authentication.

  parameters:
    analysisId:
      name: analysisId
      in: path
      required: true
      schema:
        type: string
        format: uuid
      description: The UUID of the analysis

  schemas:
    CreateAnalysisRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
          description: Human-readable analysis name (unique per user)
          example: housing-price-prediction
        description:
          type: string
          description: Optional description of the analysis
          example: Predicting California housing prices using regression models

    AnalysisInfo:
      type: object
      required: [id, name, createdAt]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
          example: housing-price-prediction
        description:
          type: string
          example: Predicting California housing prices using regression models
        createdAt:
          type: string
          format: date-time

    ComputeProfile:
      type: object
      required: [id, name, cpuRequest, cpuLimit, memoryRequest,
        memoryLimit]
      properties:
        id:
          type: string
          example: exploratory
        name:
          type: string
          example: Exploratory
        description:
          type: string
          example: Interactive data exploration and small experiments
        cpuRequest:
          type: string
          example: "1"
        cpuLimit:
          type: string
          example: "2"
        memoryRequest:
          type: string
          example: 2Gi
        memoryLimit:
          type: string
          example: 4Gi
        gpuLimit:
          type: integer
          example: 0

    LaunchWorkspaceRequest:
      type: object
      properties:
        profile:
          type: string
          description: >
            Compute profile ID. Defaults to "exploratory" if not
            specified.
          example: exploratory

    WorkspaceStatus:
      type: object
      required: [status]
      properties:
        id:
          type: string
          format: uuid
        status:
          type: string
          enum: [PENDING, RUNNING, IDLE, STOPPED, FAILED]
          example: RUNNING
        profile:
          type: string
          example: exploratory
        startedAt:
          type: string
          format: date-time
        lastActivity:
          type: string
          format: date-time
        message:
          type: string
          description: >
            Human-readable message, especially useful for FAILED
            status to communicate error details.

    WorkspaceUrl:
      type: object
      required: [url]
      properties:
        url:
          type: string
          format: uri
          description: >
            Full JupyterLab URL for iframe embedding. Includes
            the user-specific path and named server segment.
          example: http://localhost:8180/user/scientist1/a1b2c3d4/lab

    ErrorResponse:
      type: object
      required: [error, message]
      properties:
        error:
          type: string
          example: Conflict
        message:
          type: string
          example: Analysis already has an active workspace
        timestamp:
          type: string
          format: date-time